Fixed some bugs in fedora-post-install

+++ fedora-post-install.sh
@@ -24,72 +24,58 @@ if [[ $EUID -ne 0 ]]; then
    exit 1
 fi
 
-PROGRESS_FILE = ~/.fedora_install_progress
-PROGRESS = "initialization"
-echo $PROGRESS > $PROGRESS_FILE
+PROGRESS_FILE=.fedora_install_progress.txt
+PROGRESS="initialization"
+if [[ ! -f "$PROGRESS_FILE" ]]; then
+	PROGRESS="initialization"
+	echo $PROGRESS > $PROGRESS_FILE
+fi
 
 
 # Check if we are using KDE or Gnome:
 # (Just for Gtk vs Qt app preferences)
 
-DESKTOP_ENV = "unknown"
-if [[ "$XDG_CURRENT_DESKTOP" == *"GNOME"* ]]; then
-	DESKTOP_ENV = "gnome"
-elif [[ "$XDG_CURRENT_DESKTOP" == *"KDE"* ]]; then
-	DESKTOP_ENV = "kde"
-fi
-
 
 # Updating the system
 function system_update() {
-	echo "This system has an NVIDIA gpu."
 	echo "Updating packages..."
-	sleep 5
 	dnf update --refresh -y
 	sleep 5
 	echo "updated" > $PROGRESS_FILE
 	echo "Updating is done."
-	echo "The computer will reboot in 10 minutes."
+	echo "The computer will reboot in 5 minutes."
 	echo "Please run the script again after the reboot to continue."
-	sleep $((10 * 60))
+	sleep $((5 * 60))
 	shutdown -r now
 }
 
 # Installing various packages I always need
 function install_packages() {
-
+	echo "Adding RPM Fusion Free and Non-Free repositories."
 	# Add RPM-Fusion Free and Non-Free:
-	dnf install https://mirrors.rpmfusion.org/free/fedora \
-		/rpmfusion-free-release-$(rpm -E %fedora).noarch.rpm \
-		https://mirrors.rpmfusion.org/nonfree/fedora/ \
-		rpmfusion-nonfree-release-$(rpm -E %fedora).noarch.rpm
-	
+	dnf install -y https://mirrors.rpmfusion.org/free/fedora/rpmfusion-free-release-$(rpm -E %fedora).noarch.rpm https://mirrors.rpmfusion.org/nonfree/fedora/rpmfusion-nonfree-release-$(rpm -E %fedora).noarch.rpm
+
+	echo "Adding VSCode repository from Microsoft."
 	# Add VSCode Repository from Microsoft:
 	rpm --import https://packages.microsoft.com/keys/microsoft.asc
-	echo -e "[code]\n \
-	name=Visual Studio Code\n \
-	baseurl=https://packages.microsoft.com/yumrepos/vscode\n \
-	enabled=1\n \
-	autorefresh=1\n \
-	type=rpm-md\n \
-	gpgcheck=1\n \
-	gpgkey=https://packages.microsoft.com/keys/microsoft.asc" \
-	| tee /etc/yum.repos.d/vscode.repo > /dev/null
-	
+	echo -e "[code]\nname=Visual Studio Code\nbaseurl=https://packages.microsoft.com/yumrepos/vscode\nenabled=1\nautorefresh=1\ntype=rpm-md\ngpgcheck=1\ngpgkey=https://packages.microsoft.com/keys/microsoft.asc" | tee /etc/yum.repos.d/vscode.repo > /dev/null
+
 
 	# Installing packages
-	
-	if [[ $DESKTOP_ENV == "kde" ]]; then
+
+	echo "Installing packages ..."
+# 	if [[ "$XDG_CURRENT_DESKTOP" == *"KDE"* ]]; then
+# 		echo "KDE desktop detected."
 		# Internet
-		dnf install google-chrome-stable
+		dnf install -y google-chrome-stable
 		# Media
-		dnf install \
+		dnf install -y \
 			vlc \
 			obs-studio \
 			kdenlive \
 			kcolorchooser
 		# Development
-		dnf install \
+		dnf install -y \
 			@development-tools \
 			code \
 			VirtualBox \
@@ -100,29 +86,31 @@ function install_packages() {
 			kate \
 			nodejs \
 			golang
-	elif [[ $DESKTOP_ENV == "gnome" ]]; then
-		# Internet
-		dnf install google-chrome-stable
-		# Media
-		dnf install \
-			obs-studio \
-			gimp \
-		# Development
-		dnf install \
-			@development-tools
-			code \
-			VirtualBox \
-			cascadia-mono-fonts \
-			java-latest-openjdk-devel \
-			code \
-			nodejs \
-			golang \
-			geany \
-			geany-plugin-* \
-			geany-themes
-	fi
+# 	elif [[ "$XDG_CURRENT_DESKTOP" == *"GNOME"* ]]; then
+# 		echo "Gnome desktop detected."
+# 		# Internet
+# 		dnf install -y google-chrome-stable
+# 		# Media
+# 		dnf install -y \
+# 			obs-studio \
+# 			gimp \
+# 		# Development
+# 		dnf install -y \
+# 			@development-tools
+# 			code \
+# 			VirtualBox \
+# 			cascadia-mono-fonts \
+# 			java-latest-openjdk-devel \
+# 			code \
+# 			nodejs \
+# 			golang \
+# 			geany \
+# 			geany-plugin-* \
+# 			geany-themes
+# 	fi
 
 	# flatpaks
+	flatpak remote-add --if-not-exists flathub https://dl.flathub.org/repo/flathub.flatpakrepo
 	flatpak install -y flathub \
 		com.discordapp.Discord \
 		com.brave.Browser
@@ -150,8 +138,8 @@ function nvidia() {
 			xorg-x11-drv-nvidia \
 			xorg-x11-drv-nvidia-libs \
 			xorg-x11-drv-nvidia-libs.i686
-	
-	
+
+
 		echo "Installation done."
 		echo "Waiting for 10 minutes for the system to load all modules."
 		sleep $((10 * 60))
@@ -166,6 +154,8 @@ function nvidia() {
 	else
 		echo "This system does not have an NVIDIA gpu."
 		echo "No gpu drivers need to be installed."
+		echo "nvidia" > $PROGRESS_FILE
+		echo "Run the script again to continue."
 	fi
 }
 
@@ -175,14 +165,14 @@ function all_done() {
 }
 
 
-PROGRESS = $(cat $PROGRESS_FILE)
+PROGRESS=$(cat $PROGRESS_FILE)
 
 if [[ $PROGRESS == "initialization" ]]; then
-	system_update()
+	system_update
 elif [[ $PROGRESS == "updated" ]]; then
-	nvidia()
+	nvidia
 elif [[ $PROGRESS == "nvidia" ]]; then
-	install_packages()
-elif [[ $PROGRESS == "packages"]]; then
-	all_done()
+	install_packages
+elif [[ $PROGRESS == "packages" ]]; then
+	all_done
 fi
